+++ tests/testthat/test-wb_styles.R
@@ -42,7 +42,8 @@ test_that("test add_fill()", {
 
 
   # check fill
-  exp <- c("<fill><patternFill patternType=\"none\"/></fill>", "<fill><patternFill patternType=\"gray125\"/></fill>",
+  exp <- c("<fill><patternFill patternType=\"none\"/></fill>",
+           "<fill><patternFill patternType=\"gray125\"/></fill>",
            "<fill><patternFill patternType=\"solid\"><fgColor rgb=\"FFFFFF00\"/></patternFill></fill>"
   )
   got <- wb$styles_mgr$styles$fills
@@ -55,7 +56,8 @@ test_that("test add_fill()", {
   wb$add_fill("S1", dims = "A1:E6", color = c(rgb = "FFFFFF00"), every_nth_col = 2)
   wb$add_fill("S1", dims = "A1:E6", color = c(rgb = "FF00FF00"), every_nth_row = 2)
 
-  exp <- c("<fill><patternFill patternType=\"none\"/></fill>", "<fill><patternFill patternType=\"gray125\"/></fill>",
+  exp <- c("<fill><patternFill patternType=\"none\"/></fill>",
+           "<fill><patternFill patternType=\"gray125\"/></fill>",
            "<fill><patternFill patternType=\"solid\"><fgColor rgb=\"FFFFFF00\"/></patternFill></fill>",
            "<fill><patternFill patternType=\"solid\"><fgColor rgb=\"FF00FF00\"/></patternFill></fill>"
   )
@@ -78,5 +80,91 @@ test_that("test add_fill()", {
   got <- wb$worksheets[[1]]$sheet_data$cc$c_s
   expect_equal(exp, got)
 
+
+})
+
+test_that("test add_font()", {
+
+  wb <- wb_workbook()
+  wb$add_worksheet("S1")$add_data("S1", mtcars)
+  expect_silent(wb$add_font("S1", dims = "A1:K1", color = c(rgb = "FFFFFF00")))
+
+  # check xf
+  exp <- c("<xf numFmtId=\"0\" fontId=\"0\" fillId=\"0\" borderId=\"0\" xfId=\"0\"/>",
+           "<xf applyFont=\"1\" borderId=\"0\" fillId=\"0\" fontId=\"1\" numFmtId=\"0\" xfId=\"0\"/>"
+  )
+  got <- wb$styles_mgr$styles$cellXfs
+
+  expect_equal(exp, got)
+
+
+  # check font
+  exp <- c("<font><sz val=\"11\"/><color rgb=\"FF000000\"/><name val=\"Calibri\"/><family val=\"2\"/><scheme val=\"minor\"/></font>",
+           "<font><color rgb=\"FFFFFF00\"/><name val=\"Calibri\"/><sz val=\"11\"/></font>"
+  )
+  got <- wb$styles_mgr$styles$fonts
+
+  expect_equal(exp, got)
+
+  # check the actual styles
+  exp <- c("1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "",
+           "", "", "", "", "", "", "", "")
+  got <- head(wb$worksheets[[1]]$sheet_data$cc$c_s, 20)
+  expect_equal(exp, got)
+
+})
+
+test_that("test add_numfmt()", {
+
+  wb <- wb_workbook()
+  wb$add_worksheet("S1")$add_data("S1", mtcars)
+  expect_silent(wb$add_numfmt("S1", dims = "A1:A33", numfmt = 2))
+  expect_silent(wb$add_numfmt("S1", dims = "F1:F33", numfmt = "#.0"))
+
+  # check xf
+  exp <- c("<xf numFmtId=\"0\" fontId=\"0\" fillId=\"0\" borderId=\"0\" xfId=\"0\"/>",
+           "<xf applyNumberFormat=\"1\" borderId=\"0\" fillId=\"0\" fontId=\"0\" numFmtId=\"2\" xfId=\"0\"/>",
+           "<xf applyNumberFormat=\"1\" borderId=\"0\" fillId=\"0\" fontId=\"0\" numFmtId=\"165\" xfId=\"0\"/>"
+  )
+  got <- wb$styles_mgr$styles$cellXfs
+
+  expect_equal(exp, got)
+
+
+  # check numfmt
+  exp <- "<numFmt numFmtId=\"165\" formatCode=\"#.0\"/>"
+  got <- wb$styles_mgr$styles$numFmts
+
+  expect_equal(exp, got)
+
+  # check the actual styles
+  exp <- c("1", "", "", "", "", "2", "", "", "", "", "", "1", "", "",
+           "", "", "2", "", "", "")
+  got <- head(wb$worksheets[[1]]$sheet_data$cc$c_s, 20)
+  expect_equal(exp, got)
+
+})
+
+test_that("test add_cell_style()", {
+
+  wb <- wb_workbook()
+  wb$add_worksheet("S1")$add_data("S1", mtcars)
+  expect_silent(wb$add_cell_style("S1", dims = "A1:A33", textRotation = "45"))
+  expect_silent(wb$add_cell_style("S1", dims = "F1:F33", horizontal = "center"))
+
+  # check xf
+  exp <- c("<xf numFmtId=\"0\" fontId=\"0\" fillId=\"0\" borderId=\"0\" xfId=\"0\"/>",
+           "<xf borderId=\"0\" fillId=\"0\" fontId=\"0\" numFmtId=\"0\" xfId=\"0\"><alignment textRotation=\"45\"/></xf>",
+           "<xf borderId=\"0\" fillId=\"0\" fontId=\"0\" numFmtId=\"0\" xfId=\"0\"><alignment horizontal=\"center\"/></xf>"
+  )
+  got <- wb$styles_mgr$styles$cellXfs
+
+  expect_equal(exp, got)
+
+  # check the actual styles
+  exp <- c("1", "", "", "", "", "2", "", "", "", "", "", "1", "", "",
+           "", "", "2", "", "", "")
+  got <- head(wb$worksheets[[1]]$sheet_data$cc$c_s, 20)
+  expect_equal(exp, got)
 
 })